feat(ui): surface data export one tap off the Self tab
Owners reported they couldn't find the export features. Export/import
lived three steps deep — Self → "Data & sources" → scroll to the third
card — behind a label that reads like wearable setup, not "take my data
out". The tools worked; nobody could reach them.

Move the export/import controls into their own DataExportScreen (route
"your_data") and add a "Your data & export" row at the top of the Data
section on the Self landing page, mirroring the existing one-tap Privacy
Dashboard pattern. DataSourcesScreen is now scoped to connections only.

No change to export behaviour — JSON/CSV/FHIR/PDF paths are relocated
verbatim.
